* commit 'f3776822cf958afa9c161d897124553e056e647f':
  DEEC-187 fix slice button after big model is loaded
  DEEC-187 fix vulnerabilities

+++ src/main/java/com/ysoft/dctrl/editor/importer/AbstractModelImporter.java
@@ -46,6 +46,6 @@ public abstract class AbstractModelImporter<R> implements ModelImporter<R> {
 
     @Override
     public double getProgress() {
-        return size != 0 ? bytesRead*100/size : 0;
+        return size != 0 ? bytesRead*100/ (double) size : 0;
     }
 }
